# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API VisionHub360",
        "description": "API RESTful con JWT y Swagger",
        "version": "1.0.0"
    },
    "paths": {
        "/api/accountsReceivable": {
            "get": {
                "tags": [
                    "Cuentas por Cobrar"
                ],
                "summary": "Listar Cuentas por Cobrar filtradas por compañía y sucursal",
                "operationId": "9aa28130a59f89dd30dcc4a27bea9316",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "billed",
                                "paid",
                                "canceled"
                            ],
                            "example": "pending"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de Cuentas por Cobrar"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/accountsReceivable/{accountReceivable}/pay": {
            "post": {
                "tags": [
                    "Cuentas por Cobrar"
                ],
                "summary": "Registrar el pago recibido de un proveedor de crédito (liquidación)",
                "description": "Actualiza el estado a 'paid' y calcula la comisión y días transcurridos.",
                "operationId": "826ea4a3352fb4974175e835b84ce2b8",
                "parameters": [
                    {
                        "name": "accountReceivable",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "paid_amount": {
                                        "description": "Monto real recibido por su empresa del proveedor.",
                                        "type": "number",
                                        "format": "float",
                                        "example": 650000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cuenta por Cobrar liquidada exitosamente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/inventory/generateBarcodes": {
            "post": {
                "tags": [
                    "Gestión de Inventario"
                ],
                "summary": "Generar un PDF con códigos de barras para ítems de inventario seleccionados.",
                "operationId": "3048492f4ba32c98f40e5b520100f19f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "items_to_generate"
                                ],
                                "properties": {
                                    "items_to_generate": {
                                        "description": "Lista de ítems de inventario con la cantidad de códigos de barras a generar para cada uno.",
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "inventory_id",
                                                "quantity"
                                            ],
                                            "properties": {
                                                "inventory_id": {
                                                    "description": "ID del ítem de inventario.",
                                                    "type": "integer",
                                                    "example": 4
                                                },
                                                "quantity": {
                                                    "description": "Número de etiquetas a generar para este ítem.",
                                                    "type": "integer",
                                                    "example": 5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "PDF generado y listo para descarga."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/companies/{id}/roles/setup": {
            "post": {
                "tags": [
                    "Roles y Permisos"
                ],
                "summary": "Inicializa roles y permisos base para una empresa",
                "operationId": "66bd5aa19cda39f5b641790a0a2919e0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la empresa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Roles y permisos creados correctamente"
                    },
                    "403": {
                        "description": "No autorizado para esta acción"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}